pyproject.toml
[tool.poetry]
name = "library-service-api"
version = "0.1.0"
description = ""
authors = ["Volodymyr Vereshchak <volodymyr.vereshchak@gmail.com>"]
readme = "README.md"
packages = [{include = "library_service_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
djangorestframework = "^3.14.0"
black = "^22.12.0"
djangorestframework-simplejwt = "^5.2.2"
python-telegram-bot = "^13.15"
python-dotenv = "^0.21.0"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
redis = "^4.4.0"
stripe = "^5.0.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
